Reworking for Guice DI, breaking apart the wiring into discrete dependencies. Still needs tidying and further work.
