+++ app/uk/gov/hmrc/selfservicetimetopay/config/frontendWiring.scala
@@ -44,18 +44,6 @@ object WSHttp extends WSGet with WSPut with WSPost with WSDelete with AppName wi
   override val hooks: Seq[HttpHook] = Seq(AuditingHook)
 }
 
-object SessionCacheConnector extends KeystoreConnector with AppName with ServicesConfig {
-  override val sessionKey: String = getConfString("keystore.sessionKey", throw new RuntimeException("Could not find session key"))
-
-  override def defaultSource: String = appName
-
-  override def baseUri: String = baseUrl("keystore")
-
-  override def domain: String = getConfString("keystore.domain", throw new RuntimeException("Could not find config keystore.domain"))
-
-  override def http: HttpGet with HttpPut with HttpDelete = WSHttp
-}
-
 object FrontendAuditConnector extends Auditing with AppName with RunMode {
   override lazy val auditingConfig = LoadAuditingConfig("auditing")
 }
@@ -64,117 +52,4 @@ object FrontendAuthConnector extends AuthConnector with ServicesConfig {
   override val serviceUrl: String = baseUrl("auth")
 
   override def http: HttpGet = WSHttp
-}
-
-object DirectDebitConnector extends DirectDebitConnector with ServicesConfig {
-  lazy val directDebitURL: String = baseUrl("direct-debit")
-  lazy val serviceURL = "direct-debit"
-  lazy val http = WSHttp
-}
-
-object CalculatorConnector extends CalculatorConnector with ServicesConfig {
-  val calculatorURL: String = baseUrl("time-to-pay-calculator")
-  val serviceURL = "time-to-pay-calculator/paymentschedule"
-  val http = WSHttp
-}
-
-object ArrangementConnector extends ArrangementConnector with ServicesConfig {
-  val arrangementURL: String = baseUrl("time-to-pay-arrangement")
-  val serviceURL = "ttparrangements"
-  val http = WSHttp
-}
-
-object TaxPayerConnector extends TaxPayerConnector with ServicesConfig {
-  val taxPayerURL: String = baseUrl("taxpayer")
-  val serviceURL = "taxpayer"
-  val http = WSHttp
-}
-
-object EligibilityConnector extends EligibilityConnector with ServicesConfig {
-  val eligibilityURL: String = baseUrl("time-to-pay-eligibility")
-  val serviceURL = "time-to-pay-eligibility/eligibility"
-  val http = WSHttp
-}
-
-trait TimeToPayController extends FrontendController with Actions with CheckSessionAction {
-  checkSessionAction: CheckSessionAction =>
-
-  override val sessionProvider: SessionProvider = new SessionProvider() {}
-  override lazy val authConnector: AuthConnector = FrontendAuthConnector
-  protected lazy val sessionCache: KeystoreConnector = SessionCacheConnector
-  protected lazy val Action: ActionBuilder[Request] = checkSessionAction andThen PlayAction
-  protected type AsyncPlayUserRequest = AuthContext => Request[AnyContent] => Future[Result]
-  protected lazy val authenticationProvider: GovernmentGateway = SaGovernmentGateway
-  protected lazy val saRegime = SaRegime(authenticationProvider)
-  private val timeToPayConfidenceLevel = new IdentityConfidencePredicate(ConfidenceLevel.L200, Future.successful(Redirect(routes.SelfServiceTimeToPayController.getUnavailable())))
-
-  def AuthorisedSaUser(body: AsyncPlayUserRequest): PlayAction[AnyContent] = AuthorisedFor(saRegime, timeToPayConfidenceLevel).async(body)
-
-  def authorizedForSsttp(block: (Option[TTPSubmission] => Future[Result]))(implicit authContext: AuthContext, hc: HeaderCarrier): Future[Result] = {
-      sessionCache.get.flatMap[Result] {
-        case Some(TTPSubmission(_, _, _, _, _, _, _, _, Some(EligibilityStatus(false, _)), _)) =>
-          Future.successful(Redirect(routes.SelfServiceTimeToPayController.getTtpCallUs()))
-        case optSubmission =>
-          block(optSubmission)
-      }
-  }
-
-  override implicit def hc(implicit request: Request[_]): HeaderCarrier = {
-    request.cookies.find(_.name == ttpSessionId).fold(super.hc(request)) { id =>
-      super.hc(request).withExtraHeaders(ttpSessionId -> id.value)
-    }
-  }
-
-  protected def updateOrCreateInCache(found: (TTPSubmission) => TTPSubmission, notFound: () => TTPSubmission)(implicit hc: HeaderCarrier) = {
-    sessionCache.get.flatMap {
-      case Some(ttpSubmission) =>
-        Logger.info("TTP data found - merging record")
-        sessionCache.put(found(ttpSubmission))
-      case None =>
-        Logger.info("No TTP Submission data found in cache")
-        sessionCache.put(notFound())
-    }
-  }
-}
-
-trait ServiceRegistry extends ServicesConfig {
-  lazy val auditConnector: Auditing = FrontendAuditConnector
-  lazy val directDebitConnector: DirectDebitConnector = DirectDebitConnector
-  lazy val sessionCacheConnector: KeystoreConnector = SessionCacheConnector
-  lazy val authConnector: AuthConnector = FrontendAuthConnector
-  lazy val arrangementConnector: ArrangementConnector = ArrangementConnector
-  lazy val eligibilityConnector: EligibilityConnector = EligibilityConnector
-  lazy val calculatorConnector: CalculatorConnector = CalculatorConnector
-  lazy val taxPayerConnector: TaxPayerConnector = TaxPayerConnector
-}
-
-trait ControllerRegistry {
-  registry: ServiceRegistry =>
-  private lazy val controllers = Map[Class[_], Controller](
-    classOf[DirectDebitController] -> new DirectDebitController(messagesApi, directDebitConnector),
-    classOf[CalculatorController] -> new CalculatorController(messagesApi, calculatorConnector),
-    classOf[ArrangementController] -> new ArrangementController(messagesApi, directDebitConnector, arrangementConnector, calculatorConnector, taxPayerConnector, eligibilityConnector),
-    classOf[EligibilityController] -> new EligibilityController(messagesApi),
-    classOf[SelfServiceTimeToPayController] -> new SelfServiceTimeToPayController(messagesApi)
-  )
-
-  def getController[A](controllerClass: Class[A]): A = controllers(controllerClass).asInstanceOf[A]
-
-  import play.api.Play.current
-  import play.api.i18n._
-
-  // lazy val messagesApi = current.injector.instanceOf[MessagesApi]  
-  // lazy val directDebitConnector = current.injector.instanceOf[DirectDebitConnector]
-  // lazy val calculatorConnector = current.injector.instanceOf[CalculatorConnector]
-  // lazy val arrangementConnector = current.injector.instanceOf[ArrangementConnector]
-  // lazy val taxPayerConnector = current.injector.instanceOf[TaxPayerConnector]
-  // lazy val eligibilityConnector = current.injector.instanceOf[EligibilityConnector]
-
-  def messagesApi: MessagesApi  
-  def directDebitConnector: DirectDebitConnector
-  def calculatorConnector: CalculatorConnector
-  def arrangementConnector: ArrangementConnector
-  def taxPayerConnector: TaxPayerConnector
-  def eligibilityConnector: EligibilityConnector  
-
 }
