+++ app/uk/gov/hmrc/selfservicetimetopay/connectors/EligibilityConnector.scala
@@ -19,9 +19,13 @@ package uk.gov.hmrc.selfservicetimetopay.connectors
 import uk.gov.hmrc.play.http.{HeaderCarrier, HttpPost}
 import uk.gov.hmrc.selfservicetimetopay.models.{EligibilityRequest, EligibilityStatus, SelfAssessment}
 import uk.gov.hmrc.selfservicetimetopay.modelsFormat._
+import uk.gov.hmrc.play.config.ServicesConfig
+import uk.gov.hmrc.selfservicetimetopay.config.WSHttp
+import com.google.inject._
 
 import scala.concurrent.Future
 
+@ImplementedBy(classOf[EligibilityConnectorImpl])
 trait EligibilityConnector {
 
   val eligibilityURL: String
@@ -31,4 +35,11 @@ trait EligibilityConnector {
   def checkEligibility(eligibilityRequest: EligibilityRequest)(implicit hc: HeaderCarrier): Future[EligibilityStatus] = {
     http.POST[EligibilityRequest, EligibilityStatus](s"$eligibilityURL/$serviceURL", eligibilityRequest)
   }
+}
+
+@Singleton
+class EligibilityConnectorImpl extends EligibilityConnector with ServicesConfig {
+  val eligibilityURL: String = baseUrl("time-to-pay-eligibility")
+  val serviceURL = "time-to-pay-eligibility/eligibility"
+  val http = WSHttp
 }
