+++ app/uk/gov/hmrc/selfservicetimetopay/controllers/CalculatorController.scala
@@ -21,7 +21,6 @@ import java.time.LocalDate
 
 import play.api.Logger
 import play.api.mvc._
-import uk.gov.hmrc.selfservicetimetopay.config.TimeToPayController
 import uk.gov.hmrc.selfservicetimetopay.connectors.CalculatorConnector
 import uk.gov.hmrc.selfservicetimetopay.forms.CalculatorForm
 import uk.gov.hmrc.selfservicetimetopay.models._
