+++ conf/application.conf
@@ -16,7 +16,7 @@ include "common.conf"
 
 appName = "self-service-time-to-pay-frontend"
 play.http.requestHandler = "play.api.http.GlobalSettingsHttpRequestHandler"
-play.modules.enabled += "uk.gov.hmrc.play.config.inject.ConfigModule"
+#play.modules.enabled += "uk.gov.hmrc.play.config.inject.ConfigModule"
 play.modules.enabled += "com.kenshoo.play.metrics.PlayModule"
 
 # Router
